fix bugs

+++ src/main/java/lol/kv3rk/draft_predict/ServerApplication/RankedSoloQ/RankedDbRequests/DTO/BestDuo.java
@@ -6,7 +6,7 @@ public interface BestDuo {
 
     String getChampion2();
 
-    double getPickRate();
+    int getPickRate();
 
     double getWinRate();
 }
